Move AlertDialogContext to context folder
