+++ apps/web/src/app/_components/alert-dialog/hooks/use-alert-dialog.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { AlertDialogContext } from '../alert-dialog-provider';
+import { AlertDialogContext } from '../contexts/alert-dialog-context';
 
 export const useAlertDialog = () => {
   return useContext(AlertDialogContext);
